+++ R/data_adolescent_society.R
@@ -1,0 +1,13 @@
+#' One-mode subset of adolescent society dataset
+#'
+#' @docType data
+#' @keywords datasets
+#' @name adolescent_society
+#' @usage data(adolescent_society)
+#' @format tidygraph graph object
+#' @references Coleman, James S. 1961. The Adolescent Society. 
+#' New York:Free Press.
+#' 
+#' Feld, Scott. 1991. “Why your friends have more friends than you do”
+#' American Journal of Sociology 96(6): 1464-1477.
+"adolescent_society"